Port parity tooling to iced 0.14

+++ crates/pinta-theme/src/tokens/spacing.rs
@@ -1,22 +1,22 @@
 #[derive(Debug, Clone)]
 pub struct Spacing {
-    pub xxs: u16,
-    pub xs: u16,
-    pub sm: u16,
-    pub md: u16,
-    pub lg: u16,
-    pub xl: u16,
+    pub xxs: f32,
+    pub xs: f32,
+    pub sm: f32,
+    pub md: f32,
+    pub lg: f32,
+    pub xl: f32,
 }
 
 impl Default for Spacing {
     fn default() -> Self {
         Self {
-            xxs: 2,
-            xs: 4,
-            sm: 6,
-            md: 8,
-            lg: 12,
-            xl: 16,
+            xxs: 2.0,
+            xs: 4.0,
+            sm: 6.0,
+            md: 8.0,
+            lg: 12.0,
+            xl: 16.0,
         }
     }
 }
